Code refactoring FIX -task 2.3 corrected -
