+++ src/main/java/com/example/projectmanager/model/entity/Retailer.java
@@ -3,8 +3,6 @@ package com.example.projectmanager.model.entity;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import lombok.ToString;
-
 import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -17,7 +15,6 @@ import javax.persistence.JoinTable;
 import javax.persistence.ManyToMany;
 import javax.persistence.Table;
 import java.util.HashSet;
-import java.util.Objects;
 import java.util.Set;
 
 @Entity
@@ -34,26 +31,11 @@ public class Retailer {
     @Column(name = "name")
     private String name;
 
-
     @ManyToMany(cascade = {CascadeType.ALL}, fetch = FetchType.LAZY)
-    @ToString.Exclude
     @JoinTable(
             name = "retailer_product",
             joinColumns = {@JoinColumn(name = "id_retailer")},
             inverseJoinColumns = {@JoinColumn(name = "id_product")}
     )
     private Set<Product> products = new HashSet<>();
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Retailer retailer = (Retailer) o;
-        return id.equals(retailer.id) && name.equals(retailer.name);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id, name);
-    }
 }
