+++ src/main/java/com/example/projectmanager/model/entity/enums/RetailerName.java
@@ -1,0 +1,6 @@
+package com.example.projectmanager.model.entity.enums;
+
+public enum RetailerName {
+    RET_A,
+    RET_B
+}
