--- src/main/java/com/example/projectmanager/model/entity/retName/RetailerName.java
@@ -1,7 +1,0 @@
-package com.example.projectmanager.model.entity.retName;
-
-public enum RetailerName {
-
-    RET_A,
-    RET_B
-}
