+++ src/main/java/com/example/projectmanager/service/ProductService.java
@@ -1,16 +1,13 @@
 package com.example.projectmanager.service;
 
 import com.example.projectmanager.model.entity.Product;
-
 import java.util.List;
 
 public interface ProductService {
 
     void insertRandomProducts(Integer quantity);
-
     void incrementStockLevelByRetailerName(String name);
 
     List<Product> findByKeyWord(String keyword);
-
-    List<Product> findAll();
+    List<Product> findAllProducts();
 }
